requirements.txt
numpy
torch
torchvision
opencv-python
Pillow
tqdm
matplotlib
einops
lpips
accelerate
transformers
huggingface-hub
packaging
diffusers
timm
